Move Range to module.
